core/serializers.py
from rest_framework.serializers import ModelSerializer

from core.models import Categoria, Marca

class CategoriaSerializer(ModelSerializer):
    class Meta:
        model = Categoria
        fields = "__all__"


class MarcaSerializer(ModelSerializer):
    class Meta:
        model = Marca
        fields = "__all__"