pyproject.toml
[tool.poetry]
name = "garagem"
version = "0.1.0"
description = ""
authors = ["Beatriz Cristina de Arruda <biatsmax@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
